moves shd into src/scripts folder
moves PCA into src folder
